Load comments from an AJAX JSON request
Fire off an ajax request to get a JSON list of comments.

Note how we now access `data` through `this.state` rather than
`this.props`, and how `data` is initialised inside the new
`getInitialState` handler.

When the AJAX request is successful it calls `setState` with the new
data. React *automatically* re-renders when the state changes.

Also note that `componentWillMount` is a React called before rendering.